Safari-611.1.10/LayoutTests/crypto/subtle/hkdf-import-key-derive-hmac-key-expected.txt

Test deriving HMAC Keys with imported HKDF base key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS derivedKey.type is 'secret'
PASS derivedKey.extractable is false
PASS derivedKey.algorithm.name is 'HMAC'
PASS derivedKey.algorithm.length is 512
PASS derivedKey.algorithm.hash.name is 'SHA-1'
PASS derivedKey.usages is ['sign', 'verify']
PASS successfullyParsed is true

TEST COMPLETE
